// src/app/shared/models/models.ts
export interface SubMenu {
  menuValue: string;
  route: string;
  base: string;
 
}
export interface MenuItem {
  menuValue: string;
  hasSubRoute: boolean;
  showSubRoute: boolean;
  base: string;
  route?: string;
  img?: string;
  icon?: string;
  faIcon?: boolean;
  subMenus: SubMenu[];
  roles?: string[];

}

export interface SubMenu {
  menuValue: string;
  route: string;
  base: string;
  roles?: string[];
}

export interface SideBarData {
  tittle: string;
  showAsTab: boolean;
  separateRoute: boolean;
  menu: MenuItem[];
}